refactor(utils, rarity-section): ♻️ update imports and interfaces for card grouping
- Moved type definitions (CardGroup, CardCountMap) to `rarity-section.tsx`
- Updated `groupSingleCardById` function in `group-single-card-by-id.ts` to use new imports
- Adjusted `rarity-section.tsx` to reflect updated type imports and interfaces for better organization
